clean: Rename option check_length to enable_checks
This makes the option name more future proof in case the plugin will
perform more validations besides checking the length at some point.
